+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog
 
 ### Bug fixes
 
+* Guard against non-string metadata map keys
+  [#790](https://bugsnag.atlassian.net/browse/PLAT-4685)
+
 * Quiet some Analyzer false positives
   [#789](https://github.com/bugsnag/bugsnag-cocoa/pull/789)